TODO: 수정시에 input value에 data 값 집어넣기 -> 일차증가 버튼 조건문

+++ audiro/src/main/java/com/audiro/repository/TravelPlanDao.java
@@ -11,7 +11,7 @@ public interface TravelPlanDao {
 
 	Long insertTravelPlan(@Param("plan") TravelPlan plan);
 
-	int insertDetailedPlan(DetailedPlan detailedPlan);
+	Long insertDetailedPlan(DetailedPlan detailedPlan);
 
 	List<TravelPlan> selectAllTrvaelPlan();
